finish time() global++ every 1 min and add milisec everytime call time()

+++ tasks/libc/include/unistd.h
@@ -14,7 +14,7 @@
 
 ssize_t read(int fd, void *buf, size_t count);
 ssize_t write(int fd, const void *buf, size_t count);
-unsigned long time(void);
+uint32_t time(void);
 void sleep(unsigned long millis);
 void period(int, void (*f)(void *), void *);
